AudioRecording page 컴포넌트 분리하여 코드 정리
